refactoring to mocks
