+++ app/client/__tests__/setup/setupTests.js
@@ -4,26 +4,7 @@ import Adapter from "enzyme-adapter-react-16";
 
 configure({ adapter: new Adapter() });
 
-class LocalStorageMock {
-  constructor() {
-    this.store = {};
-  }
+import "../../__mocks__/local-storage-mock.js";
+import mockStore from "../../__mocks__/redux-mock-store.js";
 
-  clear() {
-    this.store = {};
-  }
-
-  getItem(key) {
-    return this.store[key] || null;
-  }
-
-  setItem(key, value) {
-    this.store[key] = value.toString();
-  }
-
-  removeItem(key) {
-    delete this.store[key];
-  }
-}
-
-global.localStorage = new LocalStorageMock();
+jest.mock("../../createStore.js", () => mockStore);
