+++ app/client/app.jsx
@@ -2,48 +2,11 @@ import "babel-polyfill";
 import "react-dates/initialize";
 
 import React from "react";
-import { createStore, applyMiddleware, compose } from "redux";
-import { connect, Provider } from "react-redux";
-import createSagaMiddleware from "redux-saga";
-import { delayConfiguration } from "pusher-redux";
-import * as Sentry from "@sentry/browser";
+import { Provider } from "react-redux";
 import { createBrowserHistory } from "history";
-import { version } from "../package.json";
 
-import appReducer from "./reducers/rootReducer";
-import rootsaga from "./sagas/rootSaga";
-
+import store from "./createStore.js";
 import Nav from "./nav.jsx";
-
-const sagaMiddleware = createSagaMiddleware();
-
-// Setup redux dev tools
-const composeSetup =
-  process.env.NODE_ENV !== "prod" &&
-  typeof window === "object" &&
-  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    : compose;
-
-export const store = createStore(
-  appReducer,
-  composeSetup(applyMiddleware(sagaMiddleware))
-);
-
-// Setup sentry
-Sentry.init({
-  dsn: window.SENTRY_REACT_DSN,
-  release: "sempo-blockchain-react@" + version
-});
-
-// Pusher Options
-const pusherOptions = {
-  cluster: "ap1"
-};
-
-delayConfiguration(store, window.PUSHER_KEY, pusherOptions);
-
-sagaMiddleware.run(rootsaga);
 
 export const browserHistory = createBrowserHistory({});
