+++ app/package.json
@@ -31,8 +31,8 @@
     "@babel/cli": "^7.1.2",
     "@babel/core": "^7.1.2",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-react": "^7.7.4",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
     "@types/chart.js": "^2.8.9",
     "@types/jest": "^25.1.4",
     "@types/node": "^13.1.4",
